Add executable-driven CLI corpus coverage and unify command tests
