+++ cli/cli_bal_run_test.go
@@ -17,28 +17,19 @@
 package main
 
 import (
-	"bytes"
-	"errors"
 	"flag"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"runtime"
 	"sort"
 	"strings"
 	"testing"
-
-	"golang.org/x/tools/txtar"
-)
-
-const (
-	coverDirEnv = "BAL_GOCOVERDIR"
 )
 
 var update = flag.Bool("update", false, "update CLI test outputs")
 
-func TestSingleBalFile(t *testing.T) {
+func TestBalRunCorpus(t *testing.T) {
 	if runtime.GOOS == "js" || runtime.GOARCH == "wasm" {
 		t.Skip("skipping CLI integration test on WASM (js/wasm)")
 	}
@@ -51,7 +42,7 @@ func TestSingleBalFile(t *testing.T) {
 	coverDir := getCoverageDir(t)
 	balBin := buildBalBinary(t, repoRoot, coverDir)
 	testDataRoot := filepath.Join(repoRoot, "corpus", "cli", "testdata")
-	outputsRoot := filepath.Join(repoRoot, "corpus", "cli", "outputs")
+	outputsRoot := filepath.Join(repoRoot, "corpus", "cli", "run")
 
 	singleBalFiles := listPaths(t, filepath.Join(testDataRoot, "single-bal-files"), true)
 	projects := listPaths(t, filepath.Join(testDataRoot, "projects"), false)
@@ -65,41 +56,6 @@ func TestSingleBalFile(t *testing.T) {
 		rel := filepath.Join("projects", filepath.Base(projectDir))
 		runAndValidateCase(t, balBin, repoRoot, coverDir, outputsRoot, projectDir, rel)
 	}
-}
-
-func getCoverageDir(t *testing.T) string {
-	t.Helper()
-	coverDir, ok := os.LookupEnv(coverDirEnv)
-	if !ok || coverDir == "" {
-		return ""
-	}
-	if err := os.MkdirAll(coverDir, 0o755); err != nil {
-		t.Fatalf("failed to create %s %q: %v", coverDirEnv, coverDir, err)
-	}
-	return coverDir
-}
-
-func buildBalBinary(t *testing.T, repoRoot, coverDir string) string {
-	t.Helper()
-	tmp := t.TempDir()
-	balName := "bal"
-	if runtime.GOOS == "windows" {
-		balName = "bal.exe"
-	}
-	balBin := filepath.Join(tmp, balName)
-
-	args := []string{"build", "-o", balBin}
-	if coverDir != "" {
-		args = append(args, "-cover", "-coverpkg=./...")
-	}
-	args = append(args, "./cli/cmd")
-
-	cmd := exec.Command("go", args...)
-	cmd.Dir = repoRoot
-	if out, err := cmd.CombinedOutput(); err != nil {
-		t.Fatalf("failed to build bal binary: %v\n%s", err, string(out))
-	}
-	return balBin
 }
 
 func listPaths(t *testing.T, dir string, balFilesOnly bool) []string {
@@ -148,94 +104,4 @@ func runAndValidateCase(t *testing.T, balBin, repoRoot, coverDir, outputsRoot, r
 			)
 		}
 	})
-}
-
-func runBalCommand(t *testing.T, balBin, runPath, repoRoot, coverDir string) (stdout, stderr string, exitCode int) {
-	t.Helper()
-	cmd := exec.Command(balBin, "run", runPath)
-	cmd.Dir = repoRoot
-	if coverDir != "" {
-		cmd.Env = append(os.Environ(), "GOCOVERDIR="+coverDir)
-	}
-
-	var stdoutBuf, stderrBuf bytes.Buffer
-	cmd.Stdout = &stdoutBuf
-	cmd.Stderr = &stderrBuf
-
-	err := cmd.Run()
-	if err == nil {
-		return stdoutBuf.String(), stderrBuf.String(), 0
-	}
-	var exitErr *exec.ExitError
-	if errors.As(err, &exitErr) {
-		return stdoutBuf.String(), stderrBuf.String(), exitErr.ExitCode()
-	}
-	t.Fatalf("failed to run bal command: %v", err)
-	return "", "", 1
-}
-
-func normalizeNewlines(s string) string {
-	s = strings.ReplaceAll(s, "\r\n", "\n")
-	return strings.TrimRight(s, "\n")
-}
-
-func readExpectedTxtar(t *testing.T, txtarPath string) (stdout, stderr, exitCode string) {
-	t.Helper()
-	archive, err := txtar.ParseFile(txtarPath)
-	if err != nil {
-		t.Fatalf("failed to parse txtar file %s: %v", txtarPath, err)
-	}
-	stdoutFound, stderrFound, exitCodeFound := false, false, false
-	for _, file := range archive.Files {
-		switch file.Name {
-		case "stdout":
-			stdout = normalizeNewlines(string(file.Data))
-			stdoutFound = true
-		case "stderr":
-			stderr = normalizeNewlines(string(file.Data))
-			stderrFound = true
-		case "exitcode":
-			exitCode = normalizeNewlines(string(file.Data))
-			exitCodeFound = true
-		default:
-			t.Fatalf("unexpected file %q in %s", file.Name, txtarPath)
-		}
-	}
-	if !stdoutFound || !stderrFound || !exitCodeFound {
-		t.Fatalf("missing stdout/stderr/exitcode entries in %s", txtarPath)
-	}
-	return stdout, stderr, exitCode
-}
-
-func updateOutputArchive(t *testing.T, expectedPath, stdout, stderr, exitCode string) {
-	t.Helper()
-	archive := &txtar.Archive{
-		Files: []txtar.File{
-			{Name: "stdout", Data: []byte(stdout)},
-			{Name: "stderr", Data: []byte(stderr)},
-			{Name: "exitcode", Data: []byte(exitCode)},
-		},
-	}
-	content := txtar.Format(archive)
-	if err := os.MkdirAll(filepath.Dir(expectedPath), 0o755); err != nil {
-		t.Fatalf("failed to create output directory for %s: %v", expectedPath, err)
-	}
-	if err := os.WriteFile(expectedPath, content, 0o644); err != nil {
-		t.Fatalf("failed to write output archive %s: %v", expectedPath, err)
-	}
-}
-
-func formatExpectedGot(expected, got string) string {
-	const indent = "\t"
-	format := func(s string) string {
-		if s == "" {
-			return indent + "(empty)"
-		}
-		lines := strings.Split(s, "\n")
-		for i := range lines {
-			lines[i] = indent + lines[i]
-		}
-		return strings.Join(lines, "\n")
-	}
-	return fmt.Sprintf("expected:\n%s\n\ngot:\n%s", format(expected), format(got))
 }
