+++ corpus/cli/completion/completion-bash.txtar
@@ -1,0 +1,6 @@
+-- stdout --
+__start_bal()
+complete -o default -F __start_bal bal
+-- stderr --
+-- exitcode --
+0
