+++ corpus/cli/run/projects/panic.txtar
