+++ corpus/cli/run/projects/run-and-print.txtar
