+++ corpus/cli/run/projects/syntax-error.txtar
