+++ corpus/cli/run/single-bal-files/panic.txtar
