+++ corpus/cli/run/single-bal-files/syntax-error.txtar
