+++ corpus/cli/version/version.txtar
@@ -1,0 +1,5 @@
+-- stdout --
+Ballerina Version:
+-- stderr --
+-- exitcode --
+0
